Rename auto-updates file to updates
